🎉 **Add postgresql-18 app**
Add a new PostgreSQL instance for 'postgresql-18' to host databases.

The PostgreSQL Helm chart usually only has one instance. For specific setups where a particular version of PostgreSQL is required, it's sometimes necessary to override the default chart name and references to use a specific version. This commit addresses that by updating the chart name to `postgresql` and also updating the associated PostgreSQL configurations to reflect this change. This ensures that the correct version of PostgreSQL is deployed, along with the necessary configurations for the target databases.