feat: Update database hostnames for multiple applications
This commit updates the database hostnames for several applications to point to `postgresql-18-hl` instead of `postgresql-hl`.
This change is necessary because a new PostgreSQL instance (version 18) has been deployed, and applications need to be reconfigured to use it.
Additionally, new backup and restore commands for PostgreSQL have been added to the Makefile.